backend/stock_prediction/stock_prediction/views.py
# backend/stock_prediction/stock_prediction/views.py

from rest_framework.response import Response
from rest_framework.decorators import api_view, authentication_classes, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework_simplejwt.authentication import JWTAuthentication

@api_view(['GET'])
@authentication_classes([JWTAuthentication])
@permission_classes([IsAuthenticated])
def my_protected_view(request):
    return Response({"message": "Esta vista está protegida."})